项目重构中

+++ KSWebRTC/Script/Chat/WebRTC/RTCManager/KSSession.h
@@ -10,9 +10,9 @@
 
 @interface KSSession : NSObject
 
-@property (nonatomic,assign) int       peerId;//被叫ID
-@property (nonatomic,assign) BOOL      isCalled;//被叫
-@property (nonatomic,copy  ) NSString  *session_id;
-@property (nonatomic,copy  ) NSString  *room;
+@property (nonatomic,assign) long long     peerId;//被叫ID
+@property (nonatomic,assign) BOOL     isCalled;//被叫
+@property (nonatomic,copy  ) NSString *session_id;
+@property (nonatomic,copy  ) NSString *room;
 
 @end
